refactor: separate types into files
