+++ src/app/error/[code]/page.tsx
@@ -2,7 +2,7 @@ import { notFound } from "next/navigation";
 
 import { ErrorMessage } from "@/components/error-message";
 import { ERROR_CODES } from "@/config/constants";
-import type { ErrorCode } from "@/lib/types";
+import type { ErrorCode } from "@/types/app";
 
 const isErrorCode = (code: string | number): code is ErrorCode =>
   (ERROR_CODES[code as ErrorCode] as string | null) != null;
