+++ src/app/guide_articles/edit/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { API_URL } from "@/config/constants";
-import type { GuideArticle } from "@/lib/types";
+import type { GuideArticle } from "@/types/app";
 
 import { Editor } from "../../editor";
