+++ src/app/student_organizations/page.tsx
@@ -1,6 +1,6 @@
 import { AbstractList } from "@/components/abstract-list";
 import { API_URL } from "@/config/constants";
-import type { StudentOrganization } from "@/lib/types";
+import type { StudentOrganization } from "@/types/app";
 
 async function fetchStudentOrganizations(page: number, resultsPerPage: number) {
   try {
