+++ src/lib/enums.ts
@@ -1,0 +1,20 @@
+export enum OrganizationType {
+  ScientificClub = "scientific_club",
+  StudentOrganization = "student_organization",
+  StudentMedium = "student_medium",
+  CultureAgenda = "culture_agenda",
+  StudentCouncil = "student_council",
+}
+
+export enum OrganizationSource {
+  StudentDepartment = "student_department",
+  Manual = "manual",
+  PwrActive = "pwr_active",
+}
+
+export enum OrganizationStatus {
+  Active = "active",
+  Inactive = "inactive",
+  Dissolved = "dissolved",
+  Unknown = "unknown",
+}
