+++ src/middleware.ts
@@ -4,7 +4,7 @@ import { NextResponse } from "next/server";
 import { AUTH_STATE_COOKIE_NAME } from "@/config/constants";
 import { getCookieOptions, parseAuthCookie } from "@/lib/cookies";
 import { fetchQuery } from "@/lib/fetch-utils";
-import type { User } from "@/lib/types";
+import type { User } from "@/types/api";
 
 const REQUIRED_ROUTE_PERMISSIONS: Record<string, string[] | undefined> = {
   "/login": [],
